pyproject.toml
[tool.poetry]
name = "rectory-apps"
version = "0.0.0"
description = ""
authors = ["Adam Peacock <adam@thepeacock.net>"]

[tool.poetry.dependencies]
python = "^3.10"
django = "~=3.2"
django-environ = "*"
django-admin-sortable2 = "*"
psycopg2 = "*"
whitenoise = "*"
uvicorn = "*"
gunicorn = "*"
django-safemigrate = "*"
django-bootstrap4 = "*"
google-auth = "*"
requests = "*"
reportlab = "*"
django-bootstrap-breadcrumbs = "*"
django-anymail = {extras = ["mailgun"], version = "*"}
django-lb-health-check = "^1.0"
django-quick-jobs = "^0.2.0"
htmlBuilder = "^1.0.0"
structlog = "^21.5.0"
django-stubs-ext = "^0.4.0"

[tool.poetry.dev-dependencies]
autopep8 = "*"
rope = "*"
pylint = "*"
pytest = "*"
pytest-django = "*"
django-debug-toolbar = "*"
ipython = "*"
pylint-django = "*"
pyyaml = "*"
uritemplate = "*"
mypy = "^0.950"
black = "^22.3.0"
django-stubs = "^1.11.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
plugins = ["mypy_django_plugin.main"]

[tool.django-stubs]
django_settings_module = "core.settings"